refactor cron
